moved rendering to GameWindow class
